# docs/api.yaml
openapi: 3.0.1
info:
  title: pokerdice API
  description: pokerdice API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authCookie:
      type: apiKey
      in: cookie
      name: AuthCookie
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        total: { type: integer, minimum: 0, description: Total de elementos (ignora paginação) }
        limit: { type: integer, minimum: 0 }
        skip: { type: integer, minimum: 0 }
        next: { type: string, format: uri, nullable: true }
        prev: { type: string, format: uri, nullable: true }
        traceId: { type: string, description: ID de correlação para debugging/observabilidade }
    Problem:
      type: object
      required: [type]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }

    # USERS
    UserCreateInputModel:
      type: object
      required: [name, password, email]
      properties:
        name: { type: string }
        password: { type: string }
        email: { type: string, format: email }
        invitationCode: { type: string, nullable: true }
    UserLoginInputModel:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    UserOutputModel:
      type: object
      properties:
        userId: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        balance: { type: number, format: float }
    UserListOutputModel:
      type: array
      items: { $ref: '#/components/schemas/UserOutputModel' }
    UserSessionOutputModel:
      type: object
      required: [userId, username]
      properties:
        userId: { type: integer }
        username: { type: string }
        token: { type: string, description: "Incluído se também suportarem header Authorization" }

    # LOBBIES
...
paths:
...
  /api/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserLoginInputModel' }
            examples:
              alice:
                summary: LOGIN ALICE
                value: { username: "Alice", password: "teste" }
              bob:
                summary: LOGIN BOB
                value: { username: "Bob", password: "teste" }
              carol:
                summary: LOGIN CAROL
                value: { username: "Carol", password: "teste" }
              dave:
                summary: LOGIN DAVE
                value: { username: "Dave", password: "teste" }
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              description: AuthCookie=...; HttpOnly; Secure; SameSite=Strict
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UserSessionOutputModel' }
              examples:
                success:
                  value:
                    data:
                      userId: 2
                      username: Bob
                      token: n3V3do3WCLY3YTmNt5ZotOUzjMlUzh78UHFyOF6D8Rk=
                    meta: null
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/users:
    get:
      tags: [Users]
      summary: List users
      operationId: fetchAllUsers
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 0 }
        - in: query
          name: skip
          schema: { type: integer, minimum: 0 }
        - in: query
          name: sort
          schema: { type: string, enum: [asc, desc] }
          description: Ordenação por nome
        - in: query
          name: contains
          schema: { type: string }
          description: Filtra por substring no nome
        - in: query
          name: name
          schema: { type: string }
          description: Filtro exato pelo nome name=Alice
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UserListOutputModel' }
                      meta: { $ref: '#/components/schemas/Meta' }
              examples:
                users:
                  value:
                    data:
                      - userId: 1
                        username: Alice
                        email: alice@example.com
                        balance: 10.0
                      - userId: 2
                        username: Bob
                        email: bob@example.com
                        balance: 8.5
                    meta:
                      total: 2
                      limit: 2
                      skip: 0
                      next: null
                      prev: null
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateInputModel' }
            examples:
              example:
                value:
                  name: "Alice"
                  email: "alice@example.com"
                  password: "teste"
                  invitationCode: "ABCD-1234"
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URL do novo utilizador
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      operationId: fetchUserById
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
              examples:
                ok:
                  value:
                    data:
                      userId: 1
                      username: Alice
                      email: alice@example.com
                      balance: 10.0
                    meta: null
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  ############################################################
  # LOBBIES
  ############################################################
...